requirements.txt
# ---- torch ----
torch==2.6.0+cu118
torchvision==0.21.0+cu118
torchaudio==2.6.0
--extra-index-url https://download.pytorch.org/whl/cu118

# ---- Transformers & HuggingFace  ----
transformers==4.52.3
accelerate==0.30.1
datasets==2.19.1
huggingface_hub==0.30.1

# ---- LoRA  ----
peft==0.10.0
bitsandbytes==0.43.1

# ---- base ----
numpy==1.26.4
scipy
tqdm
pillow
matplotlib
scikit-learn


opencv-python
soundfile
